Restructure repository into separate packages
- Create mmm-data-generator package with minimal dependencies
- Create mmm-param-recovery package with full benchmarking dependencies
- Update all import statements to use new package structure
- Update tests to use mmm_data_generator imports
- Configure pixi workspace for both packages
- Add separate READMEs for each package
- Remove old mmm_param_recovery directory structure

This allows users to install just the lightweight data generator
or the full benchmarking suite as needed.
