+++ .pixi/config.toml
@@ -1,0 +1,1 @@
+run-post-link-scripts = "insecure"
