--- mmm_param_recovery.egg-info/PKG-INFO
@@ -1,12 +1,0 @@
-Metadata-Version: 2.4
-Name: mmm-param-recovery
-Version: 0.1.0
-Summary: A package for generating synthetic MMM (Media Mix Modeling) datasets for parameter recovery studies
-Author-email: Thomas Wiecki <thomas.wiecki@gmail.com>
-Requires-Python: >=3.8
-Requires-Dist: pymc-marketing<0.15,>=0.14.0
-Requires-Dist: meridian<0.5,>=0.4.0
-Provides-Extra: dev
-Requires-Dist: pytest<9,>=8.4.1; extra == "dev"
-Requires-Dist: ruff>=0.3.0; extra == "dev"
-Requires-Dist: pre-commit>=3.0.0; extra == "dev"
