--- mmm_param_recovery.egg-info/SOURCES.txt
@@ -1,23 +1,0 @@
-pyproject.toml
-mmm_param_recovery/__init__.py
-mmm_param_recovery.egg-info/PKG-INFO
-mmm_param_recovery.egg-info/SOURCES.txt
-mmm_param_recovery.egg-info/dependency_links.txt
-mmm_param_recovery.egg-info/requires.txt
-mmm_param_recovery.egg-info/top_level.txt
-mmm_param_recovery/data_generator/__init__.py
-mmm_param_recovery/data_generator/channels.py
-mmm_param_recovery/data_generator/config.py
-mmm_param_recovery/data_generator/core.py
-mmm_param_recovery/data_generator/ground_truth.py
-mmm_param_recovery/data_generator/presets.py
-mmm_param_recovery/data_generator/regions.py
-mmm_param_recovery/data_generator/transforms.py
-mmm_param_recovery/data_generator/utils.py
-mmm_param_recovery/data_generator/validation.py
-mmm_param_recovery/data_generator/visualization.py
-tests/test_default_config.py
-tests/test_ground_truth.py
-tests/test_presets_validation.py
-tests/test_schema.py
-tests/test_transforms.py
