--- mmm_param_recovery.egg-info/requires.txt
@@ -1,7 +1,0 @@
-pymc-marketing<0.15,>=0.14.0
-meridian<0.5,>=0.4.0
-
-[dev]
-pytest<9,>=8.4.1
-ruff>=0.3.0
-pre-commit>=3.0.0
