+++ packages/mmm-data-generator/pyproject.toml
@@ -1,0 +1,67 @@
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "mmm-data-generator"
+version = "0.1.0"
+description = "A lightweight package for generating synthetic MMM (Media Mix Modeling) datasets with known ground truth parameters"
+authors = [{name = "PyMC Labs", email = "info@pymc-labs.com"}]
+readme = "README.md"
+requires-python = ">=3.10"
+license = {text = "Apache-2.0"}
+keywords = ["marketing-mix-modeling", "mmm", "synthetic-data", "parameter-recovery", "benchmarking"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: Apache Software License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Topic :: Scientific/Engineering",
+]
+
+dependencies = [
+    "numpy>=1.24.0",
+    "pandas>=2.0.0",
+    "matplotlib>=3.7.0",
+    "seaborn>=0.12.0",
+    "pymc-marketing>=0.16.0,<0.17",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.0.0",
+    "ruff>=0.1.0",
+    "pre-commit>=3.0.0",
+]
+
+[project.urls]
+Homepage = "https://github.com/pymc-labs/mmm-param-recovery"
+Documentation = "https://github.com/pymc-labs/mmm-param-recovery/tree/main/packages/mmm-data-generator"
+Repository = "https://github.com/pymc-labs/mmm-param-recovery"
+Issues = "https://github.com/pymc-labs/mmm-param-recovery/issues"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+include = ["mmm_data_generator*"]
+
+[tool.setuptools.package-data]
+mmm_data_generator = ["py.typed"]
+
+[tool.pytest.ini_options]
+testpaths = ["../../tests"]
+python_files = ["test_*.py"]
+python_classes = ["Test*"]
+python_functions = ["test_*"]
+addopts = "-v --tb=short"
+
+[tool.ruff]
+line-length = 100
+target-version = "py310"
+
+[tool.ruff.lint]
+select = ["E", "F", "W", "I", "N", "UP"]
+ignore = ["E501"]  # Line too long (handled by formatter)
+
